Factor out MoctIterTtb
